requirements.txt
webdav4[http]
pyyaml
fastapi
uvicorn
psutil
python-multipart
pyjwt
